fix accidental quotes, also renames tests
