+++ hedy.py
@@ -639,6 +639,8 @@ def process_variable(name, lookup):
     #processes a variable by hashing and escaping when needed
     if is_variable(name, lookup):
         return hash_var(name)
+    elif is_quoted(name): #sometimes kids accidentally quote strings, then we do not want them quoted again
+        return f"{name}"
     else:
         return f"'{name}'"
