+++ tests/test_level_09.py
@@ -1,6 +1,6 @@
 import hedy
 import textwrap
-from tests_level_01 import HedyTester
+from test_level_01 import HedyTester
 
 class TestsLevel9(HedyTester):
   level = 9
